+++ routes/authRoute.js
@@ -2,8 +2,9 @@ const express = require("express");
 const router = express.Router();
 
 const authController = require("../applications/auth/authController");
+const registerController = require("../applications/auth/registerController");
 
-router.post("/register", authController.register);
+router.post("/register", registerController.register);
 router.post("/login", authController.login);
 router.post("/logout", authController.logout);
